+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and the versioning aim to respect [Semantic Versioning](http://semver.org/spec/v
 
 Here is a template for new release sections
 
+## [Unreleased]
+### Added
+- multiple basemaps possible
+
 ## [1.4.1] - 2024-04-08
 ### Added
 - initial cluster docs